proxy: Split `destination` into submodules
The `destination` module is quite crowded.

In order to separate concerns:
- `Endpoint` is moved into its own module `endpoint` module;
- `Background`, `DiscoveryWork`, and `DestinationSet` have been moved
   into a dedicated `background` module;
- `Background` has been renamed to `background::Config`;
- `DiscoveryWork` has been renamed to `background::Process`;
- `destination::Watch` has been renamed `destination::Resolution`
